aider: Update welcome page to display content from welcome.yaml file correctly.
# Aider chat conversation:

USER: This works. However, the /welcome page only displays a link: Start Survey

ASSISTANT: It appears that the welcome page is not displaying the content as expected. I will provide a patch for the `welcome.html.j2` template to ensure that the content from the `welcome.yaml` file is displayed correctly on the welcome page.

Here is the diff for the change:

```diff
--- src/experiments/questionnaires/templates/welcome.html.j2
+++ src/experiments/questionnaires/templates/welcome.html.j2
@@ ... @@
 <body>
     <h1>{{ content.title }}</h1>
     <p>{{ content.introduction }}</p>
     <p>{{ content.privacy_info }}</p>
-    <p>{{ instructions }}</p>
-    <p>{{ hydration_info }}</p>
-    <p>{{ questionnaire_info }}</p>
+    <p>{{ content.instructions }}</p>
+    <p>{{ content.hydration_info }}</p>
+    <p>{{ content.questionnaire_info }}</p>
     <a href="{{ url_for('questionnaire_handler', scale='general') }}" class="start-button">Start Survey</a>
 </body>
 </html>
```

This change will ensure that the content for the instructions, hydration information, and questionnaire information is correctly pulled from the `welcome.yaml` file and displayed on the welcome page.